% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepDir.R
\name{grepDir}
\alias{grepDir}
\title{Recursive grep}
\usage{
grepDir(pattern, path, recursive = FALSE, ...)
}
\arguments{
\item{pattern}{character string containing a regular expression}

\item{path}{path to search}

\item{recursive}{logical. Should the listing recurse into directories? passed to list.files}

\item{...}{arguments passed to grep}
}
\value{
grepDir(value = FALSE) returns a vector of the indices of the elements of x that yielded a match (or not, for invert = TRUE. 
This will be an integer vector unless the input is a long vector, when it will be a double vector.
grepDir(value = TRUE) returns a character vector containing the selected elements of x (after coercion, preserving names but no other attributes).
}
\description{
Recursive call to grep in R
}
