% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFilesGithub.R
\name{list.files_github}
\alias{list.files_github}
\title{List files of a github repository subdirectory}
\usage{
list.files_github(repo, subdir = "R")
}
\arguments{
\item{repo}{character Repository address in the format username/repo}

\item{subdir}{subdirectory within repo that contains the files to list}
}
\description{
List full raw paths to files in a github repository subdirectory
}
\examples{
l=list.files_github('tidyverse/ggplot2')
l
#geom-boxplot.r
readLines(l[42])
}
