% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_github.R
\name{ls_github}
\alias{ls_github}
\title{List files of a github repository subdirectory}
\usage{
ls_github(repo, ref = "master", subdir = NULL, recursive = FALSE)
}
\arguments{
\item{repo}{character Repository address in the format username/repo}

\item{subdir}{subdirectory within repo that contains the files to list}
}
\description{
List full raw paths to files in a github repository subdirectory
}
\examples{
l=ls_github('tidyverse/ggplot2')
l
#geom-boxplot.r
readLines(l[42])
}
