% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeImport.R
\name{makeImport}
\alias{makeImport}
\title{Scrape R script to create namespace calls}
\usage{
makeImport(file, cut = NULL, print = FALSE, format = "oxygen")
}
\arguments{
\item{file}{character connection to pass to readLines, can be file path, directory path, url path}

\item{cut}{integer number of functions to write as importFrom until switches to import}

\item{print}{boolean print output to console, default FALSE}

\item{format}{character the output format must be in c('oxygen','namespace','description'), default oxygen}
}
\description{
Scrape r script to create namespace calls for roxygen, namespace or description files
}
\examples{
makeImport(list.files_github('yonicd/YSmisc','R'),print = T,format = 'oxygen')
makeImport(list.files_github('yonicd/YSmisc','R'),print = T,format = 'namespace')
makeImport(list.files_github('yonicd/YSmisc','R'),print = T,format = 'description')
}
