% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tween_factor.R
\name{tween_factor}
\alias{tween_factor}
\title{FUNCTION_TITLE}
\usage{
tween_factor(data, levels, direction.mat, ...)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{levels}{PARAM_DESCRIPTION}

\item{direction.mat}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[plyr]{ddply}}
 \code{\link[tweenr]{tween_states}}
}
\author{
Jonathan Sidi
}
