% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorised_centroid.R
\name{vectorised.centroid}
\alias{vectorised.centroid}
\title{FUNCTION_TITLE}
\usage{
vectorised.centroid(polygon)
}
\arguments{
\item{polygon}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Jonathan Sidi
}
