% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_repo.R
\name{show_repo}
\alias{show_repo}
\title{Visually inspect structure of git repository before initial fetch}
\usage{
show_repo(layout = "collapse", update_archive = FALSE)
}
\arguments{
\item{layout}{charaacter, Layout of d3Tree output collapse, cartesian, radial Default: 'collapse'}

\item{update_archive}{boolean, update the archive file, Default: FALSE}
}
\value{
nothing
}
\description{
After adding a remote visually inpsect the directory structure of it before fetching/pulling
}
\seealso{
\code{\link[d3Tree]{d3tree}}
}
