% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_repo.R
\name{show_repo}
\alias{show_repo}
\title{Visually inspect structure of git repository before initial fetch}
\usage{
show_repo(path = getwd(), layout = "collapse")
}
\arguments{
\item{path}{character, Path to root directory of git repository or a name of a github repository, Default: setwd()}

\item{layout}{character, Layout of d3Tree output collapse, cartesian, radial Default: 'collapse'}
}
\value{
data.frame
}
\description{
Visually inpsect the directory structure of it before cloning/fetching/pulling
}
\details{
By default path assumes a local address, if path is a valid repository name eg 'tidyverse/glue' 
then the pattern is checked on the specified pulic github repository
}
\examples{
show_repo('tidyverse/glue')
}
\seealso{
\code{\link[d3Tree]{d3tree}}
}
